fastest a_star
