+++ example/android/.gitignore
@@ -9,5 +9,3 @@ GeneratedPluginRegistrant.java
 # Remember to never publicly share your keystore.
 # See https://flutter.dev/docs/deployment/android#reference-the-keystore-from-the-app
 key.properties
-**/*.keystore
-**/*.jks
